rename semester file to calendar
