+++ packages/time/src/browser.ts
@@ -3,7 +3,7 @@ import cpf from 'dayjs/plugin/customParseFormat.js';
 
 dayjs.extend(cpf);
 
-export * as semester from './semester.js';
+export * as semester from './calendar.js';
 export * as tzOffset from './tz-offset.js';
 export {default as unitsToMs} from './units-to-ms.js';
